Allow removal of disconnected Jetpack purchases
Due the UI preventing any management of purchases if the user has no sites, an edge case
happens where a user who has no sites, made a Jetpack purchase and then disconnected
that Jetpack purchase is unable to manage their purchase. This will move the check that
sites exist to after the call to fetch user purchases, allowing all purchases to be shown,
regardless of the number of sites to which a user belongs. For listing purchases, if there
are no purchases found, it will show the no sites message. Managing purchases have the
check for user sites removed entirely as manage purchases will redirect to /me/purchases if
something is wrong.