Remove siteSelection middleware in cancelPurchase
The siteSelection middleware in the route that allows users to cancel
purchases does not work for disconnected Jetpack purchases for users
who belong to no other sites. When this is present, the route is never
hit because siteSelect requires a site to proceed and because the
Jetpack purchase was disconnected, the user is no longer a user on that
site. In this case, siteSelection was causing a "No Sites" message to
be displayed.