Allow Purchase Cancellation without Site
This commit clears the path for a user with a disconnected Jetpack
purchase and no other sites (i.e., when the user only purchased Jetpack
for their .org site) to actually cancel and refund their purchase. This
requires a few changes to the code path that verify that a user has
sites or that a site is selected.