Set selectedSite to false when missing
For disconnected Jetpack users who have no other purchases, some
components will not have access to a selectedSite, because no site will
be found (disconnecting Jetpack removes the user from the site). This
requires updating the getSelectedSite function to return false instead
of null and updating the propTypes to accept either a bool or object
for the selectedSite prop.