Use siteSlug and siteId in place of selectedSite
Only two properties were ever used from the selectedSite object in this
component tree: ID and slug. Components were refactored to allow for
no selectedSite object to be required, so users with only disconnected
Jetpack purchases in their account could still cancel and refund their
purchases. This also disables eslint for some warnings about classes
which are outside of the scope of this PR and not trivial to fix.